Make sure htdocs is readable
This fixes a problem on servers where the umask is set such that the files downloaded via git do not get world-readable/listable permissions. Without this change, the permissions from the working directory propagate into the portal container and apache is not able to read the data on servers with umask 007.