refactor: menu: remove fixup component for link/menu
Remove the fixup components MenuItemForLink and DropDownItemForLink and put <Route> elements in <Switch> in an array.

(The fixup-components where introduced to suppress a warning that was triggered if an <Menu.Item> was put in a <Link>. If we put the <Route> elements in an array instead of an <>-element, the warning is gone and we can remove the fixup elements.

Solution was suggested here: (https://stackoverflow.com/questions/66463284/react-router-warning-route-elements-should-not-change-from-controlled-to-unco)